reworked fix to not use Path.GetExtension
