+++ src/Sarif.Sdk.sln
@@ -120,8 +120,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sarif.Multitool.Library", "
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Test.UnitTests.Sarif.Multitool", "Test.UnitTests.Sarif.Multitool\Test.UnitTests.Sarif.Multitool.csproj", "{C90C5EA4-765A-4D32-A96F-2D15F14FDB2B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Test.UnitTests.Sarif.Driver.Net48", "Test.UnitTests.Sarif.Driver.Net48\Test.UnitTests.Sarif.Driver.Net48.csproj", "{8DCA3EDD-033B-4792-B5FB-A3D89660B670}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -348,18 +346,6 @@ Global
 		{C90C5EA4-765A-4D32-A96F-2D15F14FDB2B}.Release|x64.Build.0 = Release|Any CPU
 		{C90C5EA4-765A-4D32-A96F-2D15F14FDB2B}.Release|x86.ActiveCfg = Release|Any CPU
 		{C90C5EA4-765A-4D32-A96F-2D15F14FDB2B}.Release|x86.Build.0 = Release|Any CPU
-		{8DCA3EDD-033B-4792-B5FB-A3D89660B670}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8DCA3EDD-033B-4792-B5FB-A3D89660B670}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8DCA3EDD-033B-4792-B5FB-A3D89660B670}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{8DCA3EDD-033B-4792-B5FB-A3D89660B670}.Debug|x64.Build.0 = Debug|Any CPU
-		{8DCA3EDD-033B-4792-B5FB-A3D89660B670}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{8DCA3EDD-033B-4792-B5FB-A3D89660B670}.Debug|x86.Build.0 = Debug|Any CPU
-		{8DCA3EDD-033B-4792-B5FB-A3D89660B670}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8DCA3EDD-033B-4792-B5FB-A3D89660B670}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8DCA3EDD-033B-4792-B5FB-A3D89660B670}.Release|x64.ActiveCfg = Release|Any CPU
-		{8DCA3EDD-033B-4792-B5FB-A3D89660B670}.Release|x64.Build.0 = Release|Any CPU
-		{8DCA3EDD-033B-4792-B5FB-A3D89660B670}.Release|x86.ActiveCfg = Release|Any CPU
-		{8DCA3EDD-033B-4792-B5FB-A3D89660B670}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
