--- src/Test.UnitTests.Sarif.Driver.Net48/AnalyzeTestContext.cs
@@ -1,14 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using Microsoft.CodeAnalysis.Sarif;
-
-namespace Test.UnitTests.Sarif.Driver.Net48
-{
-    public class AnalyzeTestContext : AnalyzeContextBase
-    {
-        public override void Dispose()
-        {
-        }
-    }
-}
