--- src/Test.UnitTests.Sarif.Driver.Net48/MSTestSettings.cs
@@ -1,4 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-[assembly: Parallelize(Scope = ExecutionScope.MethodLevel)]
